+++ spec/support/contentful_helpers.rb
@@ -49,6 +49,7 @@ module ContentfulHelpers
       options: hash_response.dig("fields", "options"),
       type: hash_response.dig("fields", "type"),
       next: double(id: hash_response.dig("fields", "next", "sys", "id")),
+      primary_call_to_action: hash_response.dig("fields", "primaryCallToAction"),
       raw: raw_response,
       content_type: double(id: hash_response.dig("sys", "contentType", "sys", "id"))
     )